chore: refactor burger constructor item
