+++ src/features/burger-constructor/componets/index.ts
@@ -1,1 +1,1 @@
-export * from './empty-item'
+export * from './burger-constructor-item'
